Websites Scraping All Words
This Python code defines a web scraping spider using the Scrapy framework to scrape data from multiple websites. The spider is named "multi_website." It starts by specifying a list of website URLs in the start_urls variable, which the spider will crawl.

For each URL, the spider sends an HTTP request to the website, extracts the text content from paragraphs and the title, preprocesses the text by converting it to lowercase, removing non-alphabetical characters, and eliminating common stopwords. The code also calculates word frequencies in the text and performs sentiment analysis on the filtered text using VADER sentiment analysis.

The spider stores the scraped data, including the website URL, filtered text content, and sentiment score, in a dictionary and yields this data as it crawls through the website's pages. It also follows links on each page, continuing to scrape data from linked pages, up to a maximum of 200 links.

After crawling all the websites and pages, the spider generates a CSV file named 'multi_website_word_frequencies_sentiment.csv.' This file includes information about word frequencies and sentiment scores for each word encountered during the crawl, along with the website associated with each word. This data can be useful for text analysis and sentiment analysis across multiple websites.